refactor: Move to rag_modules
