+++ src/rag_modules/chat_model.py
