+++ src/rag_modules/embedding_model.py
